requirements.txt
python-telegram-bot[ext]==13.15
feedparser==6.0.10
requests==2.31.0
